* src/gnome/window-main.c: Kevin Finn's patch to shutdown gnucash 	when the main window is deleted. Handle remove_child events to 	prevent accessing deleted windows.
	* src/engine/sixtp-dom-parsers.c (string_to_integer): same as
	below

	* src/engine/sixtp-dom-generators.c (add_kvp_value_node): same as
	below

	* src/engine/sixtp-xml-write-utils.c: same as below

	* src/engine/sixtp-utils.c: same as below

	* src/engine/gnc-numeric.c: same as below

	* src/engine/TransLog.c: same as below

	* src/engine/kvp_frame.c: same as below

	* src/engine/sql/kvp-sql.c: same as below

	* src/engine/sql/builder.c (sqlBuild_Set_Int64): don't assume
	long long int == gint64

	* src/gnome/window-main.c: take out devel menu

	* src/gnome/dialog-price-editor.c (price_ok_clicked): fix bug


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@3976 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/engine/TransLog.c
@@ -224,10 +224,10 @@ xaccTransWriteLog (Transaction *trans, char flag)
                split->memo ? split->memo : "",
                split->action ? split->action : "",
                split->reconciled,
-               gnc_numeric_num(split->damount), 
-               gnc_numeric_denom(split->damount),
-               gnc_numeric_num(split->value), 
-               gnc_numeric_denom(split->value),
+               (long long int) gnc_numeric_num(split->damount), 
+               (long long int) gnc_numeric_denom(split->damount),
+               (long long int) gnc_numeric_num(split->value), 
+               (long long int) gnc_numeric_denom(split->value),
                drecn ? drecn : "");
 
       g_free (drecn);
